Move bundle parser into its own file
