+++ impeller/compiler/shader_bundle.h
@@ -1,0 +1,13 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "impeller/compiler/switches.h"
+
+namespace impeller {
+namespace compiler {
+
+bool GenerateShaderBundle(Switches& switches);
+
+}  // namespace compiler
+}  // namespace impeller
